+++ Sources/SwifterSwift/SwiftStdlib/FloatingPointExtensions.swift
@@ -1,12 +1,5 @@
 // FloatingPointExtensions.swift - Copyright 2025 SwifterSwift
 
-#if canImport(Foundation)
-import Foundation
-#endif
-#if canImport(Android)
-import Android
-#endif
-
 // MARK: - Properties
 
 public extension FloatingPoint {
@@ -25,56 +18,26 @@ public extension FloatingPoint {
         return self < 0
     }
 
-    #if canImport(Foundation) && !os(Android)
     /// SwifterSwift: Ceil of number.
     var ceil: Self {
-        return Foundation.ceil(self)
+        return rounded(.up)
     }
-    #endif
+
+    /// SwifterSwift: Floor of number.
+    var floor: Self {
+        return rounded(.down)
+    }
 
     /// SwifterSwift: Radian value of degree input.
     var degreesToRadians: Self {
         return Self.pi * self / Self(180)
     }
 
-    #if canImport(Foundation) && !os(Android)
-    /// SwifterSwift: Floor of number.
-    var floor: Self {
-        return Foundation.floor(self)
-    }
-    #endif
-
     /// SwifterSwift: Degree value of radian input.
     var radiansToDegrees: Self {
         return self * Self(180) / Self.pi
     }
 }
-
-#if os(Android)
-public extension Double {
-    /// SwifterSwift: Ceil of number.
-    var ceil: Self {
-        return Foundation.ceil(self)
-    }
-
-    /// SwifterSwift: Floor of number.
-    var floor: Self {
-        return Foundation.floor(self)
-    }
-}
-
-public extension Float {
-    /// SwifterSwift: Ceil of number.
-    var ceil: Self {
-        return Foundation.ceilf(self)
-    }
-
-    /// SwifterSwift: Floor of number.
-    var floor: Self {
-        return Foundation.floorf(self)
-    }
-}
-#endif
 
 // MARK: - Operators
 
@@ -114,7 +77,7 @@ prefix operator √
 /// - Returns: square root of given float.
 public prefix func √ <T>(float: T) -> T where T: FloatingPoint {
     // http://nshipster.com/swift-operators/
-    return sqrt(float)
+    return float.squareRoot()
 }
 
 // swiftlint:enable identifier_name