+++ Sources/SwifterSwift/UIKit/IBInspectable/UITextFieldInspectableExtensions.swift
@@ -1,0 +1,34 @@
+// UITextFieldInspectableExtensions.swift - Copyright 2025 SwifterSwift
+
+#if canImport(UIKit) && !os(watchOS)
+import UIKit
+
+public extension UITextField {
+    /// SwifterSwift: Left view tint color.
+    @IBInspectable var leftViewTintColor: UIColor? {
+        get {
+            guard let iconView = leftView as? UIImageView else { return nil }
+            return iconView.tintColor
+        }
+        set {
+            guard let iconView = leftView as? UIImageView else { return }
+            iconView.image = iconView.image?.withRenderingMode(.alwaysTemplate)
+            iconView.tintColor = newValue
+        }
+    }
+
+    /// SwifterSwift: Right view tint color.
+    @IBInspectable var rightViewTintColor: UIColor? {
+        get {
+            guard let iconView = rightView as? UIImageView else { return nil }
+            return iconView.tintColor
+        }
+        set {
+            guard let iconView = rightView as? UIImageView else { return }
+            iconView.image = iconView.image?.withRenderingMode(.alwaysTemplate)
+            iconView.tintColor = newValue
+        }
+    }
+}
+
+#endif