+++ Sources/SwifterSwift/UIKit/UIViewExtensions.swift
@@ -763,7 +763,7 @@ public extension UIView {
     ///   - color: The separator color. Default is `.black`.
     ///   - height: The height of the separator. Default is `1`.
     ///   - spacing: Spacing from the bottom edge. Default is `0`.
-    func addViewSeparator(color: UIColor = .black, height: CGFloat = 1, spacing: CGFloat = 0) {
+    func addBottomSeparator(color: UIColor = .black, height: CGFloat = 1, spacing: CGFloat = 0) {
         let line = UIView()
         line.translatesAutoresizingMaskIntoConstraints = false
         line.backgroundColor = color