+++ Tests/UIKitTests/UIViewExtensionsTests.swift
@@ -684,23 +684,19 @@ final class UIViewExtensionsTests: XCTestCase { // swiftlint:disable:this type_b
         XCTAssertFalse(view.subviews.first is UIVisualEffectView)
     }
     
-   
-    func testAddViewSeparator() throws {
+    func testAddBottomSeparator() throws {
         let view = UIView()
         let expectedColor: UIColor = .red
         let expectedHeight: CGFloat = 2.0
 
-        view.addViewSeparator(color: expectedColor, height: expectedHeight)
+        view.addBottomSeparator(color: expectedColor, height: expectedHeight)
 
         let separator = try XCTUnwrap(view.subviews.first, "Separator not found")
-        
+
         XCTAssertEqual(view.subviews.count, 1)
         XCTAssertEqual(separator.backgroundColor, expectedColor)
-        XCTAssertEqual(separator.frame.height, expectedHeight, accuracy: 0.1)
+        XCTAssertEqual(separator.frame.height, expectedHeight)
     }
-
-    
-    
 }
 
 #endif