+++ CHANGELOG.md
@@ -2,15 +2,15 @@
 
 The changelog for **SwifterSwift**. Also see the [releases](https://github.com/SwifterSwift/SwifterSwift/releases) on GitHub.
 
-## [Unreleased]
+
+
+## Upcoming Release
 
 ### Added
 
 - **UIView**
   - Added `separator(color:height:spacing:)` method to easily insert a bottom line to a view. 🚀 [#1234](https://github.com/SwifterSwift/SwifterSwift/pull/1234) by [sangjin](https://github.com/SsangG77)
 
-
-## Upcoming Release
 ### Added
 - **Android**
   - Added Android platform support. [#1230](https://github.com/SwifterSwift/SwifterSwift/pull/1230) by [Marc Prud'hommeaux](https://github.com/marcprux)